FIX: Add io.iohk prefix to packages. (#16)
