+++ metronome/core/src/io/iohk/metronome/core/package.scala
@@ -1,0 +1,5 @@
+package io.iohk.metronome
+
+package object core {
+  type Validated[U] = Validated.Tagged[U]
+}
