--- metronome/core/src/metronome/core/package.scala
@@ -1,5 +1,0 @@
-package metronome
-
-package object core {
-  type Validated[U] = Validated.Tagged[U]
-}
