+++ metronome/crypto/src/io/iohk/metronome/crypto/hash/Hash.scala
@@ -1,0 +1,6 @@
+package io.iohk.metronome.crypto.hash
+
+import io.iohk.metronome.core.Tagger
+import scodec.bits.ByteVector
+
+object Hash extends Tagger[ByteVector]
