+++ metronome/crypto/src/io/iohk/metronome/crypto/hash/package.scala
@@ -1,0 +1,5 @@
+package io.iohk.metronome.crypto
+
+package object hash {
+  type Hash = Hash.Tagged
+}
