--- metronome/crypto/src/metronome/crypto/GroupSignature.scala
@@ -1,7 +1,0 @@
-package metronome.crypto
-
-/** Group signature of members with identity `K` over some content `H`,
-  * represented by type `G`, e.g. `G` could be a `List[Secp256k1Signature]`
-  * or a single combined threshold signature of some sort.
-  */
-case class GroupSignature[K, H, G](sig: G)
