--- metronome/crypto/src/metronome/crypto/hash/Hash.scala
@@ -1,6 +1,0 @@
-package metronome.crypto.hash
-
-import metronome.core.Tagger
-import scodec.bits.ByteVector
-
-object Hash extends Tagger[ByteVector]
