--- metronome/crypto/src/metronome/crypto/hash/package.scala
@@ -1,5 +1,0 @@
-package metronome.crypto
-
-package object hash {
-  type Hash = Hash.Tagged
-}
