+++ metronome/hotstuff/consensus/src/io/iohk/metronome/hotstuff/consensus/ViewNumber.scala
@@ -1,0 +1,13 @@
+package io.iohk.metronome.hotstuff.consensus
+
+import io.iohk.metronome.core.Tagger
+
+object ViewNumber extends Tagger[Long] {
+  implicit class Ops(val vn: ViewNumber) extends AnyVal {
+    def next: ViewNumber = ViewNumber(vn + 1)
+    def prev: ViewNumber = ViewNumber(vn - 1)
+  }
+
+  implicit val ord: Ordering[ViewNumber] =
+    Ordering.by(identity[Long])
+}
