+++ metronome/hotstuff/consensus/src/io/iohk/metronome/hotstuff/consensus/package.scala
@@ -1,0 +1,5 @@
+package io.iohk.metronome.hotstuff
+
+package object consensus {
+  type ViewNumber = ViewNumber.Tagged
+}
