--- metronome/hotstuff/consensus/src/metronome/hotstuff/consensus/package.scala
@@ -1,5 +1,0 @@
-package metronome.hotstuff
-
-package object consensus {
-  type ViewNumber = ViewNumber.Tagged
-}
